start sharing common code between runner and validator
